+++ askanna/cli/tool.py
@@ -1,9 +1,11 @@
 # -*- coding: utf-8 -*-
-from askanna.cli.utils import init_checks
-from askanna.cli.utils import update_available
-import askanna
-import click
 import importlib
+import sys
+
+import askanna
+from askanna.core.utils import init_checks
+from askanna.core.utils import update_available
+import click
 from dotenv import find_dotenv, load_dotenv
 load_dotenv(find_dotenv())
 
@@ -23,10 +25,21 @@ For usage and help on a specific command, run it with a --help flag, e.g.:
 CONTEXT_SETTINGS = {'help_option_names': ['-h', '--help']}
 
 
+def cli():
+    """
+    This is a wrapper to catch any command and errors
+    """
+    try:
+        cli_commands()
+    except Exception as e:
+        click.echo(e)
+        sys.exit(1)
+
+
 @click.group(help=HELP, short_help=SHORT_HELP, epilog=EPILOG,
              context_settings=CONTEXT_SETTINGS)
 @click.version_option(askanna.__version__)
-def cli():
+def cli_commands():
     update_url = update_available()
     if update_url:
         click.echo("INFO: A newer version of AskAnna is available. Update "
@@ -45,13 +58,14 @@ commands = [
     "artifact",
     "upload_result",
     "variable",
+    "run",
 ]
 
 for command in commands:
     module_path = "askanna.cli." + command
     command_module = importlib.import_module(module_path)
     command_name = command.replace('_', '-')
-    cli.add_command(command_module.cli, command_name)
+    cli_commands.add_command(command_module.cli, command_name)
 
 
 # perform initial checks on config and auth status