+++ askanna/core/exceptions.py
@@ -1,0 +1,14 @@
+class Error(Exception):
+    pass
+
+
+class GetError(Error):
+    pass
+
+
+class PostError(Error):
+    pass
+
+
+class RunError(Error):
+    pass