+++ tests/base.py
@@ -2,14 +2,14 @@ import unittest
 
 from click.testing import CliRunner
 
-from askanna.cli.tool import cli
+from askanna.cli.tool import cli_commands
 
 
 class BaseCLItest(unittest.TestCase):
     verb = ""
 
     def setUp(self):
-        self.result = CliRunner().invoke(cli, ['--help', self.verb, ])
+        self.result = CliRunner().invoke(cli_commands, ['--help', self.verb, ])
 
     def testCommandLineAccess(self):
         assert self.result.exit_code == 0